Create Make Factor
Chef has two integers A and B (A ≤ B).Chef can choose any non-negative integer X and add them to both A and B. Find whether it is possible to make A a divisor of B.
Input:
The only line of input contains two integers A and B.

Constraints:
1 ≤ A ≤ B ≤ 109
Output:
Return "Yes" if it is possible to make A a factor of B, "No" otherwise.
Example:
Sample Input:
3 6
Sample Output:
Yes
Explanation:
We can choose X=0 and add them to 3 and 6. Thus, 3 is a factor of 6.