Create Make Sum of Array Zero
Chef has an array A of length N consisting of 1 and −1 only. In one operation, Chef can choose any index i such that 1 ≤ i ≤N and multiply the element Ai​ by −1. Find the minimum number of operations required to make the sum of the array equal to 0. Return-1 if the sum of the array cannot be made 0.

Input:
The first line of input contains N denoting the length of array A.
The following line denotes the N space separated integers of the array A.

Constraints:
2 ≤ N ≤ 1000
Ai = 1 or Ai = -1
Output:
Return the minimum number of operations required to make the sum of the array equal to 0. Return -1 if the sum of the array cannot be made 0.
Example:
Sample Input:
4
1 1 1 1
Sample Output:
2
Explanation:
The minimum number of operations required is 2. In the first operation, change A3 from 1 to -1. Similarly in the second operation change A4 from 1 to -1. Thus the sum of final array is 1+1-1-1 =0