Create Perfect Trio
Varun defines a group of three friends as a perfect group if the age of one person is equal to the sum of the age of remaining two people.
Given that, the ages of three people in a group are A, B, and C respectively, find whether the group is perfect.

Input
The first line of input will contain a single integer T, denoting the number of test cases.
Each test case consists of three space-separated integers A,B, and C - the age of the three members of group.

Constraints
1 ≤ T ≤ 200
1 ≤ A, B, C ≤ 100
Output
For each test case, output on a new line, Yes, if the group is perfect, and No otherwise.

Example
Sample Input
4
10 20 30
23 51 17
Sample Output
Yes
No
Explanation
Test case 1: The sum of age of first and second person is equal to that of the third person. Thus, the group is perfect.
Test case 2: There exists no member in the group such that the age of one person is equal to the sum of the age of remaining two people. Thus, the group is not perfect.